Introduce interpolated_region module
- literal text and command are parsed to a rust strings at first.
- Then, the rust string is translated to a flagment.
